requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
gunicorn>=21.2.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0